% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_chart.R
\name{draw_whole_sign_chart}
\alias{draw_whole_sign_chart}
\title{Visualize chart in whole sign stype}
\usage{
draw_whole_sign_chart(planet_position)
}
\arguments{
\item{planet_position}{a data frame (obtained by calculate_planet_position)}
}
\description{
Visualize chart in whole sign stype
}
