% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{globals}
\alias{globals}
\title{define global variables}
\format{
An object of class \code{character} of length 1.
}
\usage{
globals
}
\description{
define global variables
}
\keyword{datasets}
