% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainpage_module.R
\name{mainpage_ui}
\alias{mainpage_ui}
\title{UI part Shiny module: main page}
\usage{
mainpage_ui(id, i18n)
}
\arguments{
\item{id}{shiny module ID}

\item{i18n}{a shiny.i18n::Translator object that links different languages}
}
\description{
UI part Shiny module: main page
}
