% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_west_planets}
\alias{reorder_west_planets}
\title{Take original theta and manipulated theta, reorder to ensure planets are displayed in correct order}
\usage{
reorder_west_planets(theta, new_theta)
}
\arguments{
\item{theta}{original theta vector}

\item{new_theta}{manipulated theta vector}
}
\description{
Take original theta and manipulated theta, reorder to ensure planets are displayed in correct order
}
