% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_planet_position.R
\docType{data}
\name{SE}
\alias{SE}
\title{Swiss Ephemeris object}
\format{
An object of class \code{list} of length 217.
}
\usage{
SE
}
\description{
Swiss Ephemeris object
}
\keyword{datasets}
