% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_chart_module.R
\name{single_chart_server}
\alias{single_chart_server}
\title{Server part shiny: single chart}
\usage{
single_chart_server(id, r6)
}
\arguments{
\item{id}{Shiny module ID}

\item{r6}{R6 object to help with communication between modules (for translation etc.)}
}
\description{
Server part shiny: single chart
}
