% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_chart_module.R
\name{update_select_input_server}
\alias{update_select_input_server}
\title{Server part shiny: single chart (update SelectInput)}
\usage{
update_select_input_server(id, r6)
}
\arguments{
\item{id}{Shiny module ID}

\item{r6}{R6 object to help with communication between modules (for translation etc.)}
}
\description{
Server part shiny: single chart (update SelectInput)
}
