% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager.R
\name{DataManager}
\alias{DataManager}
\title{An R6 Class to communicate between shiny modules}
\description{
An R6 Class to communicate between shiny modules

An R6 Class to communicate between shiny modules
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{language}}{current language version}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataManager-clone}{\code{DataManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
