% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager.R
\name{DataManager}
\alias{DataManager}
\title{R6 Class to store & process data.}
\description{
An R6 Class to communicate between shiny modules
}
\section{Super class}{
\code{\link[shiny.i18n:Translator]{shiny.i18n::Translator}} -> \code{DataManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{language}}{(\code{character()})\cr
Language selected for shinyApp. Default to Chinese.}

\item{\code{horoscope_datetime}}{(\code{POSIXct()})\cr
Date & time of the horoscope. Default to current datetime.}

\item{\code{horoscope_timezone}}{(\code{character()})\cr
Time zone. Default to Asia/Taipei}

\item{\code{horoscope_city}}{(\code{character()})\cr
City with longitude & latitude. Default to Taipei City}

\item{\code{horoscope_country}}{(\code{character()})\cr
Country}

\item{\code{planet_position}}{(\code{list()})\cr
A list of dataf rame containing planetary positions and house cusps}

\item{\code{chart}}{(\code{list()})\cr
#' ggplot object}

\item{\code{chart_name}}{(\code{character()})\cr
name of the chart}

\item{\code{selected_planets}}{(\code{character()})\cr
name of the chart}

\item{\code{aspect_table}}{a data frame of aspects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataManager-new}{\code{DataManager$new()}}
\item \href{#method-DataManager-update_chart}{\code{DataManager$update_chart()}}
\item \href{#method-DataManager-clone}{\code{DataManager$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="at"><a href='../../shiny.i18n/html/Translator.html#method-Translator-at'><code>shiny.i18n::Translator$at()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="automatic_translate"><a href='../../shiny.i18n/html/Translator.html#method-Translator-automatic_translate'><code>shiny.i18n::Translator$automatic_translate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="get_key_translation"><a href='../../shiny.i18n/html/Translator.html#method-Translator-get_key_translation'><code>shiny.i18n::Translator$get_key_translation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="get_languages"><a href='../../shiny.i18n/html/Translator.html#method-Translator-get_languages'><code>shiny.i18n::Translator$get_languages()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="get_translation_language"><a href='../../shiny.i18n/html/Translator.html#method-Translator-get_translation_language'><code>shiny.i18n::Translator$get_translation_language()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="get_translations"><a href='../../shiny.i18n/html/Translator.html#method-Translator-get_translations'><code>shiny.i18n::Translator$get_translations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="parse_date"><a href='../../shiny.i18n/html/Translator.html#method-Translator-parse_date'><code>shiny.i18n::Translator$parse_date()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="parse_number"><a href='../../shiny.i18n/html/Translator.html#method-Translator-parse_number'><code>shiny.i18n::Translator$parse_number()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="set_translation_language"><a href='../../shiny.i18n/html/Translator.html#method-Translator-set_translation_language'><code>shiny.i18n::Translator$set_translation_language()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="t"><a href='../../shiny.i18n/html/Translator.html#method-Translator-t'><code>shiny.i18n::Translator$t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="translate"><a href='../../shiny.i18n/html/Translator.html#method-Translator-translate'><code>shiny.i18n::Translator$translate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny.i18n" data-topic="Translator" data-id="use_js"><a href='../../shiny.i18n/html/Translator.html#method-Translator-use_js'><code>shiny.i18n::Translator$use_js()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-new}{}}}
\subsection{Method \code{new()}}{
Initalize the R6 class using the original shiny.i18n Translator class initialize function
furthermore, initialize horoscope city, calculate planatery positions & draw charts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$new(translation_csvs_path = "./inst/csv/", separator_csv = "|")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{translation_csvs_path}}{the path to the folder where the csv files for translation are stored}

\item{\code{separator_csv}}{csv separator}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-update_chart"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-update_chart}{}}}
\subsection{Method \code{update_chart()}}{
update horoscope. Calculate planetary positions, draw chart
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$update_chart()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
