% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_chart.R
\name{date_to_jd}
\alias{date_to_jd}
\title{Convert a date/time into Julian Day}
\usage{
date_to_jd(date, hour, minute, timezone = "Asia/Taipei")
}
\arguments{
\item{date}{A string of date (ISO format) YYYY-MM-DD}

\item{hour}{An integer of hour.}

\item{minute}{An integer of minute}

\item{timezone}{A string of time zone. It has to be time zone that lubridate recognizes.}
}
\value{
An integer of Julian Day
}
\description{
Convert a date/time into Julian Day
}
