% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainpage_module.R
\name{language_pick_server}
\alias{language_pick_server}
\title{Server part shiny: language pick}
\usage{
language_pick_server(id, r6)
}
\arguments{
\item{id}{Shiny module ID}

\item{r6}{an R6 object to help with communication between modules (for translation etc.)}
}
\value{
a list with selected language in reactive function
}
\description{
Server part shiny: language pick
}
