% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optmize_planet_position}
\alias{optmize_planet_position}
\title{Optimize until all planets are not overlapping}
\usage{
optmize_planet_position(planet_theta, selected_elements)
}
\arguments{
\item{selected_elements}{all planets selected for chart}

\item{planet_thetat}{a numeric vector}
}
\description{
Optimize until all planets are not overlapping
}
